pyproject.toml
[project]
name = "ai-gym-coach-chatbot"
version = "0.1.0"
description = "FastAPI + LangChain + Gemini gym coach chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "sqlalchemy==2.0.43",
  "psycopg2-binary==2.9.10",
  "pydantic==2.11.9",
  "pydantic-settings==2.10.1",
  "python-jose[cryptography]==3.5.0",
  "passlib[bcrypt]==1.7.4",
  "python-multipart==0.0.20",
  "httpx==0.28.1",
  "langchain==1.2.7",
  "langchain-google-genai==4.2.0",
  "email-validator==2.2.0",
]

[tool.uv]
package = false
